backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=11.0.0
httpx[socks]>=0.26.0
pydantic>=2.5.3
langchain-openai>=0.3.0
google-genai>=1.0.0
pydantic-settings>=2.0.0
